requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2
django-allauth==0.48.0
django-cors-headers==3.11.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
oauth2client==4.1.3
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.6.0
djangorestframework-simplejwt==5.0.0
